+++ tests/three/adjectives/adjectives.test.js
@@ -8,7 +8,7 @@ test('adjectives misc', function (t) {
   let json = doc.adjectives().json()
   t.equal(json.length, 1, here + 'adj')
   t.equal(json[0].adjective.adverb, 'clearly', here + 'adv')
-  t.equal(json[0].adjective.noun, 'clearness', here + 'noun')
+  t.equal(json[0].adjective.noun, 'clarity', here + 'noun')
   t.equal(json[0].adjective.superlative, 'clearest', here + 'super')
   t.equal(json[0].adjective.comparative, 'clearer', here + 'comp')