+++ tests/three/api.test.js
@@ -189,7 +189,6 @@ test('api:', function (t) {
     doc.sentences().toFutureTense()
     doc.sentences().toInfinitive()
     doc.sentences().toNegative()
-    doc.sentences().debullet()
     doc.questions()
 
     // verbs